requirements.txt
duckduckgo-search
phidata
python-dotenv
groq
yfinance